Moving adding/removing logic from booking repository to service

+++ src/main/java/org/example/services/BookingService.java
@@ -7,6 +7,7 @@ import org.example.model.Booking;
 import org.example.repositories.BookingRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.Set;
 
@@ -17,11 +18,26 @@ public class BookingService {
     private final BookingRepository bookingRepository;
 
     public Booking create(final Booking booking) throws PhoneAlreadyBookedException {
-        return bookingRepository.add(booking);
+        var timestamped = booking.toBuilder().id(bookingRepository.getNextSequenceId()).createdAt(LocalDateTime.now()).build();
+        /*
+        Since only phone is used for comparing bookings, only one booking for the given phone is allowed
+         */
+        if (bookingRepository.add(timestamped)) {
+            return timestamped;
+        } else {
+            throw new PhoneAlreadyBookedException(String.format("Booking for the given phone has been already made (booking: %s)", timestamped));
+        }
     }
 
     public void cancel(final Booking booking) throws BookingNotFoundException {
-        bookingRepository.remove(booking);
+        /*
+        Since only phone is used for comparing bookings, bookings.contains(booking) would ignore a user
+         */
+        if (bookingRepository.findByPhoneAndUser(booking.getPhone(), booking.getUser()).isPresent()) {
+            bookingRepository.remove(booking);
+        } else {
+            throw new BookingNotFoundException(String.format("No booking to cancel found (booking: %s)", booking));
+        }
     }
 
     public Optional<Booking> findById(Integer id) {
